Make the project directory structure "standard"...
... where "standard" means "my best interpretation of
https://github.com/pypa/sampleproject", which I think is an authority,
and I think I got pretty close.

So I moved the test into "tests", and the "kolumnextractor" directory
into "src", and moved the "__init__.py" there.

Then I had to fiddle with pyproject to get tests running.

So, this may not be an improvement, perhaps the way things were was also
a legitimate project layout.  I'm really not a python expert.